new code

+++ src/main/java/com/openprice/chatservice/algo/NBS.java
@@ -2,7 +2,7 @@ package com.openprice.chatservice.algo;
 
 public class NBS {
 
-    public static int calculateNBS(double sellerDisagreementPoint, double buyerDisagreementPoint,
+    public static double calculateNBS(double sellerDisagreementPoint, double buyerDisagreementPoint,
             double buyerBargainingPower) {
         // Ensure that the sum of bargaining powers is equal to 1
         double sellerBargainingPower = 1 - buyerBargainingPower;
@@ -11,6 +11,6 @@ public class NBS {
         double nashBargainingSolution = sellerBargainingPower * sellerDisagreementPoint +
                 buyerBargainingPower * buyerDisagreementPoint;
 
-        return (int)Math.ceil(nashBargainingSolution);
+        return Math.ceil(nashBargainingSolution);
     }
 }
